# api-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "OrderController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The list of orders",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order was successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot create order. Invalid data provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrderController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The uuid of the order to be found.",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order with requested id.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "The order was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "put": {
        "operationId": "OrderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The uuid of the order to be updated",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order was successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "Order not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot update order. Invalid data provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The id of the order to be deleted",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order was successfully removed.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "Order not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      }
    }
  },
  "info": {
    "title": "E-Commerce Order Management API",
    "description": "RESTful API for managing users and orders in an e-commerce system",
    "version": "1.0.0",
    "contact": {
      "name": "Contact API Support",
      "url": "https://t.me/kalyonekenobe",
      "email": "oleksandr.ihumnov@ukma.edu.ua"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.ukmashop.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponseEntity": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "Invalid email"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "ApiErrorResponseEntity": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-02-15T21:24:28.191Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/endpoint"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occured on server"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null,
            "additionalProperties": {}
          },
          "errors": {
            "type": "object",
            "example": {
              "server": [
                "An error occured on server"
              ]
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "timestamp",
          "path",
          "success",
          "message",
          "data",
          "errors"
        ]
      },
...
      "PaginatedApiResponseEntity": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "Invalid email"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalCount": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "success",
          "data",
          "page",
          "limit",
          "totalCount"
        ]
      },
...
      "ProductEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the product (UUID v4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "Name of the product",
            "example": "Apple iPhone 15"
          },
          "skuNumber": {
            "type": "string",
            "description": "SKU number (Stock Keeping Unit) for inventory management",
            "example": "FS-8571954"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the product was created",
            "example": "2026-02-15T21:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the product was last updated",
            "example": "2026-02-15T21:45:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "skuNumber",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the order (UUID v4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "description": "Identifier of the user who placed the order",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "type": "string",
            "description": "Status of the order",
            "example": "Pending"
          },
          "tax": {
            "type": "string",
            "description": "Tax amount applied to the order",
            "example": "5.50"
          },
          "shippingCost": {
            "type": "string",
            "description": "Shipping cost of the order",
            "example": "10.00"
          },
          "shippingAddress": {
            "type": "string",
            "description": "Shipping address for the order",
            "example": "123 Main St, Kyiv, Ukraine"
          },
          "billingAddress": {
            "type": "string",
            "description": "Billing address for the order",
            "example": "123 Main St, Kyiv, Ukraine"
          },
          "notes": {
            "type": "object",
            "description": "Additional notes for the order",
            "example": "Leave package at the front door"
          },
          "trackingNumber": {
            "type": "object",
            "description": "Tracking number for the order shipment",
            "example": "TRACK123456789"
          },
          "estimatedDelivery": {
            "format": "date-time",
            "type": "string",
            "description": "Estimated delivery date of the order",
            "example": "2026-02-20T12:00:00Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the order was created",
            "example": "2026-02-15T21:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the order was last updated",
            "example": "2026-02-15T21:45:00Z"
          },
          "products": {
            "description": "Products included in the order with quantity and price details",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductToOrderEntity"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "status",
          "tax",
          "shippingCost",
          "shippingAddress",
          "billingAddress",
          "notes",
          "trackingNumber",
          "estimatedDelivery",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductToOrderEntity": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Identifier of the product included in the order",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "orderId": {
            "type": "string",
            "description": "Identifier of the order containing the product",
            "example": "a1b2c3d4-e5f6-7890-1234-56789abcdef0"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the product in the order",
            "example": 2
          },
          "price": {
            "type": "string",
            "description": "Price of a single unit of the product in the order",
            "example": "19.99"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the relation was created",
            "example": "2026-02-15T21:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the relation was last updated",
            "example": "2026-02-15T21:45:00Z"
          },
          "product": {
            "description": "The product included in the order",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductEntity"
              }
            ]
          },
          "order": {
            "description": "The order that includes the product",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderEntity"
              }
            ]
          }
        },
        "required": [
          "productId",
          "orderId",
          "quantity",
          "price",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrderProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID of the product being added to the order",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the product in the order",
            "example": 2
          },
          "price": {
            "type": "string",
            "description": "Price of the product at the time of order",
            "example": "15.50"
          }
        },
        "required": [
          "productId",
          "quantity",
          "price"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "shippingAddress": {
            "type": "string",
            "description": "Shipping address for the order",
            "example": "123 Main St, Kyiv, Ukraine"
          },
          "billingAddress": {
            "type": "string",
            "description": "Billing address for the order",
            "example": "123 Main St, Kyiv, Ukraine"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for the order",
            "example": "Leave package at the front door"
          },
          "estimatedDelivery": {
            "format": "date-time",
            "type": "string",
            "description": "Estimated delivery date of the order",
            "example": "2026-02-20T12:00:00Z"
          },
          "tax": {
            "type": "string",
            "description": "Tax amount applied to the order",
            "example": "5.50"
          },
          "shippingCost": {
            "type": "string",
            "description": "Shipping cost of the order",
            "example": "10.00"
          },
          "products": {
            "description": "Products included in the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderProductDto"
            }
          }
        },
        "required": [
          "shippingAddress",
          "billingAddress",
          "estimatedDelivery",
          "tax",
          "shippingCost",
          "products"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New status of the order",
            "enum": [
              "Pending",
              "Processing",
              "Shipped",
              "Delivered",
              "Cancelled"
            ],
            "example": "Shipped"
          },
          "shippingAddress": {
            "type": "string",
            "description": "New shipping address",
            "example": "456 Another St, Kyiv, Ukraine"
          },
          "billingAddress": {
            "type": "string",
            "description": "New billing address",
            "example": "456 Another St, Kyiv, Ukraine"
          },
          "trackingNumber": {
            "type": "object",
            "description": "Tracking number for the order shipment",
            "example": "TRACK123456789"
          },
          "notes": {
            "type": "string",
            "description": "Updated notes for the order",
            "example": "Leave package at the back door"
          }
        }
      }
    }
  }
}